Chore: Move brand queries to inside the brand compnent
