+++ react/components/BrandTranslation/graphql/getAllBrands.gql
