+++ react/components/BrandTranslation/graphql/getBrand.gql
