+++ react/components/BrandTranslation/graphql/translateBrand.gql
