+++ react/components/BrandTranslation/graphql/uploadBrandTranslation.gql
